10. Safe divide

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -3,9 +3,9 @@
 
 DELIMITER //
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS INT
+RETURNS FLOAT
 BEGIN
-    DECLARE result INT;
+    DECLARE result FLOAT;
     IF b = 0 THEN
         SET result = 0;
     ELSE
